refactor: rename base package from 'test' to 'finance' for better domain clarity
