+++ src/main/java/com/allobank/finance/config/OpenApiConfig.java
@@ -1,4 +1,4 @@
-package com.allobank.test.config;
+package com.allobank.finance.config;
 
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.info.Contact;
@@ -7,16 +7,8 @@ import io.swagger.v3.oas.annotations.servers.Server;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@OpenAPIDefinition(
-        info = @Info(
-                title = "Allo Backend Test API",
-                version = "v1",
-                description = "API documentation for finance data endpoints.",
-                contact = @Contact(name = "Allo Backend Candidate")
-        ),
-        servers = {
+@OpenAPIDefinition(info = @Info(title = "Allo Backend Test API", version = "v1", description = "API documentation for finance data endpoints.", contact = @Contact(name = "Allo Backend Candidate")), servers = {
                 @Server(url = "http://localhost:8080", description = "Local server")
-        }
-)
+})
 public class OpenApiConfig {
 }
