+++ src/main/java/com/allobank/finance/controller/HomeController.java
@@ -1,4 +1,4 @@
-package com.allobank.test.controller;
+package com.allobank.finance.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
