+++ src/main/java/com/allobank/finance/exception/DataNotInitializedException.java
@@ -1,4 +1,4 @@
-package com.allobank.test.exception;
+package com.allobank.finance.exception;
 
 public class DataNotInitializedException extends RuntimeException {
