+++ src/main/java/com/allobank/finance/exception/GlobalExceptionHandler.java
@@ -1,10 +1,11 @@
-package com.allobank.test.exception;
+package com.allobank.finance.exception;
 
-import com.allobank.test.service.FinanceDataService;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import com.allobank.finance.service.FinanceDataService;
 
 import java.util.Map;
