+++ src/main/java/com/allobank/finance/strategy/IDRDataFetcher.java
@@ -1,4 +1,4 @@
-package com.allobank.test.strategy;
+package com.allobank.finance.strategy;
 
 import java.util.List;
 import java.util.Map;
