+++ src/test/java/com/allobank/test/strategy/HistoricalIdrUsdFetcherTest.java
@@ -1,11 +1,13 @@
 package com.allobank.test.strategy;
 
-import com.allobank.test.client.FrankfurterClient;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+
+import com.allobank.finance.client.FrankfurterClient;
+import com.allobank.finance.strategy.HistoricalIdrUsdFetcher;
 
 import java.math.BigDecimal;
 import java.util.LinkedHashMap;
