+++ src/test/java/com/allobank/test/strategy/LatestIdrRatesFetcherTest.java
@@ -1,12 +1,14 @@
 package com.allobank.test.strategy;
 
-import com.allobank.test.client.FrankfurterClient;
-import com.allobank.test.config.FrankfurterApiProperties;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+
+import com.allobank.finance.client.FrankfurterClient;
+import com.allobank.finance.config.FrankfurterApiProperties;
+import com.allobank.finance.strategy.LatestIdrRatesFetcher;
 
 import java.math.BigDecimal;
 import java.math.MathContext;
