+++ src/test/java/com/allobank/test/strategy/SupportedCurrenciesFetcherTest.java
@@ -1,11 +1,13 @@
 package com.allobank.test.strategy;
 
-import com.allobank.test.client.FrankfurterClient;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+
+import com.allobank.finance.client.FrankfurterClient;
+import com.allobank.finance.strategy.SupportedCurrenciesFetcher;
 
 import java.util.List;
 import java.util.Map;
